refactor: Rigging python tests in pytest for CI and Nox
