+++ examples/int8/training/vgg16/requirements.txt
@@ -1,3 +1,6 @@
 torch>=1.10.0
 tensorboard>=1.14.0
-pytorch-quantization --extra-index-url https://pypi.ngc.nvidia.com
+nvidia-pyindex
+--extra-index-url https://pypi.ngc.nvidia.com
+pytorch-quantization>=2.1.2
+tqdm
