+++ tests/py/hw/test_multi_gpu.py
