+++ tests/py/ptq/test_ptq_trt_calibrator.py
@@ -8,7 +8,42 @@ import torch.nn as nn
 from torch.nn import functional as F
 import torchvision
 import torchvision.transforms as transforms
-from model_test_case import ModelTestCase
+
+def find_repo_root(max_depth=10):
+    dir_path = os.path.dirname(os.path.realpath(__file__))
+    for i in range(max_depth):
+        files = os.listdir(dir_path)
+        if "WORKSPACE" in files:
+            return dir_path
+        else:
+            dir_path = os.path.dirname(dir_path)
+
+    raise RuntimeError("Could not find repo root")
+
+MODULE_DIR = find_repo_root() + "/tests/modules"
+
+def compute_accuracy(testing_dataloader, model):
+    total = 0
+    correct = 0
+    loss = 0.0
+    class_probs = []
+    class_preds = []
+    device = torch.device('cuda:0')
+    with torch.no_grad():
+        idx = 0
+        for data, labels in testing_dataloader:
+            data, labels = data.to(device), labels.to(device)
+            out = model(data)
+            preds = torch.max(out, 1)[1]
+            class_probs.append([F.softmax(i, dim=0) for i in out])
+            class_preds.append(preds)
+            total += labels.size(0)
+            correct += (preds == labels).sum().item()
+            idx += 1
+
+    test_probs = torch.cat([torch.stack(batch) for batch in class_probs])
+    test_preds = torch.cat(class_preds)
+    return correct / total
 
 
 class TRTEntropyCalibrator(trt.IInt8EntropyCalibrator2):
@@ -54,9 +89,10 @@ class TRTEntropyCalibrator(trt.IInt8EntropyCalibrator2):
                 f.write(cache)
 
 
-class TestAccuracy(ModelTestCase):
+class TestAccuracy(unittest.TestCase):
 
-    def setUp(self):
+    def test_compile_script(self):
+        self.model = torch.jit.load(MODULE_DIR + "/trained_vgg16.jit.pt").eval().to("cuda")
         self.input = torch.randn((1, 3, 32, 32)).to("cuda")
         self.testing_dataset = torchvision.datasets.CIFAR10(root='./data',
                                                             train=False,
@@ -74,32 +110,7 @@ class TestAccuracy(ModelTestCase):
         # Test cases can assume using GPU id: 0
         self.calibrator = TRTEntropyCalibrator(self.testing_dataloader)
 
-    def compute_accuracy(self, testing_dataloader, model):
-        total = 0
-        correct = 0
-        loss = 0.0
-        class_probs = []
-        class_preds = []
-        device = torch.device('cuda:0')
-        with torch.no_grad():
-            idx = 0
-            for data, labels in testing_dataloader:
-                data, labels = data.to(device), labels.to(device)
-                out = model(data)
-                preds = torch.max(out, 1)[1]
-                class_probs.append([F.softmax(i, dim=0) for i in out])
-                class_preds.append(preds)
-                total += labels.size(0)
-                correct += (preds == labels).sum().item()
-                idx += 1
-
-        test_probs = torch.cat([torch.stack(batch) for batch in class_probs])
-        test_preds = torch.cat(class_preds)
-        return correct / total
-
-    def test_compile_script(self):
-
-        fp32_test_acc = self.compute_accuracy(self.testing_dataloader, self.model)
+        fp32_test_acc = compute_accuracy(self.testing_dataloader, self.model)
         log(Level.Info, "[Pyt FP32] Test Acc: {:.2f}%".format(100 * fp32_test_acc))
 
         compile_spec = {
@@ -116,24 +127,12 @@ class TestAccuracy(ModelTestCase):
         }
 
         trt_mod = torchtrt.ts.compile(self.model, **compile_spec)
-        int8_test_acc = self.compute_accuracy(self.testing_dataloader, trt_mod)
+        int8_test_acc = compute_accuracy(self.testing_dataloader, trt_mod)
         log(Level.Info, "[TRT INT8] Test Acc: {:.2f}%".format(100 * int8_test_acc))
         acc_diff = fp32_test_acc - int8_test_acc
         self.assertTrue(abs(acc_diff) < 3)
 
 
-def test_suite():
-    suite = unittest.TestSuite()
-    # You need a pre-trained VGG cifar10 model to run this test. Please follow instructions at
-    # https://github.com/NVIDIA/torchtrt/tree/master/cpp/ptq/training/vgg16 to export this model.
-    suite.addTest(TestAccuracy.parametrize(TestAccuracy, model=torch.jit.load('./trained_vgg16.jit.pt')))
 
-    return suite
-
-
-suite = test_suite()
-
-runner = unittest.TextTestRunner()
-result = runner.run(suite)
-
-exit(int(not result.wasSuccessful()))
+if __name__ == "__main__":
+    unittest.main()
