+++ tests/py/requirements.txt
@@ -1,2 +1,3 @@
-torchvision==0.12.0+cu113
--f https://download.pytorch.org/whl/torch_stable.html
+torchvision==0.12.0
+--extra-index-url https://download.pytorch.org/whl/cu113
+pytest
